+++ packages/components/text-field/CHANGELOG.md
@@ -3,6 +3,21 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/adevinta/spark/compare/@spark-ui/text-field@0.2.5...@spark-ui/text-field@0.3.0) (2023-06-29)
+
+### Features
+
+- **form-field:** update textfield to remove js ([6149c3e](https://github.com/adevinta/spark/commit/6149c3e18362e53e087e688735bba4785153811b))
+- **text-field:** fix typescript error ([5f0e7ad](https://github.com/adevinta/spark/commit/5f0e7ad54dd617ead6cf3e6821e90af06c5ce2e4))
+- **text-field:** remove not needed prop ([8fba304](https://github.com/adevinta/spark/commit/8fba3049f13a1ecd151c56433afc76a21b9be115))
+- **text-field:** small improvements ([bf21be8](https://github.com/adevinta/spark/commit/bf21be8bdf24b3fa0b9d0ea90fa636c889ca8e9b))
+- **text-field:** update field ([71508f5](https://github.com/adevinta/spark/commit/71508f541c20b46d590476c297373b13cd0332c4))
+- **text-field:** update label ([73b24d3](https://github.com/adevinta/spark/commit/73b24d333ba693831a26362a5c699d24ac65083d))
+- **text-field:** update styles ([e2859d8](https://github.com/adevinta/spark/commit/e2859d84719d60bf20cb5c8706f78abef2f69b6f))
+- **text-field:** update textfield and add required indicator ([6fa99c4](https://github.com/adevinta/spark/commit/6fa99c4d83afd552cfa48b0c908eea4228816f36))
+- **text-field:** use content prop ([5cb5b47](https://github.com/adevinta/spark/commit/5cb5b47eba96560b2364a12301fca16d13bf2ab8))
+- **text-field:** use input primitive ([ccda9e3](https://github.com/adevinta/spark/commit/ccda9e31f93b1bf20f57598b9662a62b2bd31130))
+
 ## [0.2.5](https://github.com/adevinta/spark/compare/@spark-ui/text-field@0.2.4...@spark-ui/text-field@0.2.5) (2023-06-26)
 
 **Note:** Version bump only for package @spark-ui/text-field